diarization_service/requirements.txt
fastapi
loguru
uvicorn
pyannote.audio
torch
pydantic